chore: format Solidity code with forge fmt
Run forge fmt to apply consistent formatting to all Solidity files:
- Update interface definitions formatting
- Improve multi-line function declarations
- Standardize spacing and indentation
- Format test files and scripts

+++ kms/auth-eth/script/Manage.s.sol
@@ -14,13 +14,13 @@ import "../contracts/DstackApp.sol";
 abstract contract BaseScript is Script {
     DstackKms public kms;
     DstackApp public app;
-    
+
     function setUp() public virtual {
         // Load from environment or use defaults
         address kmsAddr = vm.envOr("KMS_CONTRACT_ADDR", address(0));
         require(kmsAddr != address(0), "KMS_CONTRACT_ADDR not set");
         kms = DstackKms(kmsAddr);
-        
+
         // App address is optional for some scripts
         address appAddr = vm.envOr("APP_CONTRACT_ADDR", address(0));
         if (appAddr != address(0)) {
@@ -33,11 +33,11 @@ abstract contract BaseScript is Script {
 contract AddKmsAggregatedMr is BaseScript {
     function run() external {
         bytes32 mrAggregated = vm.envBytes32("MR_AGGREGATED");
-        
+
         vm.startBroadcast();
         kms.addKmsAggregatedMr(mrAggregated);
         vm.stopBroadcast();
-        
+
         console.log("Added KMS aggregated MR:", vm.toString(mrAggregated));
     }
 }
@@ -45,11 +45,11 @@ contract AddKmsAggregatedMr is BaseScript {
 contract RemoveKmsAggregatedMr is BaseScript {
     function run() external {
         bytes32 mrAggregated = vm.envBytes32("MR_AGGREGATED");
-        
+
         vm.startBroadcast();
         kms.removeKmsAggregatedMr(mrAggregated);
         vm.stopBroadcast();
-        
+
         console.log("Removed KMS aggregated MR:", vm.toString(mrAggregated));
     }
 }
@@ -57,11 +57,11 @@ contract RemoveKmsAggregatedMr is BaseScript {
 contract AddOsImage is BaseScript {
     function run() external {
         bytes32 imageHash = vm.envBytes32("OS_IMAGE_HASH");
-        
+
         vm.startBroadcast();
         kms.addOsImageHash(imageHash);
         vm.stopBroadcast();
-        
+
         console.log("Added OS image hash:", vm.toString(imageHash));
     }
 }
@@ -69,11 +69,11 @@ contract AddOsImage is BaseScript {
 contract AddKmsDevice is BaseScript {
     function run() external {
         bytes32 deviceId = vm.envBytes32("DEVICE_ID");
-        
+
         vm.startBroadcast();
         kms.addKmsDevice(deviceId);
         vm.stopBroadcast();
-        
+
         console.log("Added KMS device:", vm.toString(deviceId));
     }
 }
@@ -81,11 +81,11 @@ contract AddKmsDevice is BaseScript {
 contract SetGatewayAppId is BaseScript {
     function run() external {
         string memory gatewayId = vm.envString("GATEWAY_APP_ID");
-        
+
         vm.startBroadcast();
         kms.setGatewayAppId(gatewayId);
         vm.stopBroadcast();
-        
+
         console.log("Set gateway app ID:", gatewayId);
     }
 }
@@ -96,16 +96,13 @@ contract SetKmsInfo is BaseScript {
         bytes memory caPubkey = vm.envBytes("CA_PUBKEY");
         bytes memory quote = vm.envBytes("QUOTE");
         bytes memory eventlog = vm.envBytes("EVENTLOG");
-        
+
         vm.startBroadcast();
-        kms.setKmsInfo(DstackKms.KmsInfo({
-            k256Pubkey: k256Pubkey,
-            caPubkey: caPubkey,
-            quote: quote,
-            eventlog: eventlog
-        }));
+        kms.setKmsInfo(
+            DstackKms.KmsInfo({ k256Pubkey: k256Pubkey, caPubkey: caPubkey, quote: quote, eventlog: eventlog })
+        );
         vm.stopBroadcast();
-        
+
         console.log("KMS info set successfully");
         console.log("  K256 Pubkey length:", k256Pubkey.length);
         console.log("  CA Pubkey length:", caPubkey.length);
@@ -119,11 +116,11 @@ contract AddComposeHash is BaseScript {
     function run() external {
         require(address(app) != address(0), "APP_CONTRACT_ADDR not set");
         bytes32 composeHash = vm.envBytes32("COMPOSE_HASH");
-        
+
         vm.startBroadcast();
         app.addComposeHash(composeHash);
         vm.stopBroadcast();
-        
+
         console.log("Added compose hash:", vm.toString(composeHash));
     }
 }
@@ -132,11 +129,11 @@ contract AddDevice is BaseScript {
     function run() external {
         require(address(app) != address(0), "APP_CONTRACT_ADDR not set");
         bytes32 deviceId = vm.envBytes32("DEVICE_ID");
-        
+
         vm.startBroadcast();
         app.addDevice(deviceId);
         vm.stopBroadcast();
-        
+
         console.log("Added device:", vm.toString(deviceId));
     }
 }
@@ -145,11 +142,11 @@ contract SetAllowAnyDevice is BaseScript {
     function run() external {
         require(address(app) != address(0), "APP_CONTRACT_ADDR not set");
         bool allow = vm.envBool("ALLOW_ANY_DEVICE");
-        
+
         vm.startBroadcast();
         app.setAllowAnyDevice(allow);
         vm.stopBroadcast();
-        
+
         console.log("Set allowAnyDevice to:", allow);
     }
 }
@@ -162,17 +159,11 @@ contract DeployApp is BaseScript {
         bool allowAnyDevice = vm.envOr("ALLOW_ANY_DEVICE", true);
         bytes32 deviceId = vm.envOr("INITIAL_DEVICE_ID", bytes32(0));
         bytes32 composeHash = vm.envOr("INITIAL_COMPOSE_HASH", bytes32(0));
-        
+
         vm.startBroadcast();
-        address appAddr = kms.deployAndRegisterApp(
-            owner,
-            disableUpgrades,
-            allowAnyDevice,
-            deviceId,
-            composeHash
-        );
+        address appAddr = kms.deployAndRegisterApp(owner, disableUpgrades, allowAnyDevice, deviceId, composeHash);
         vm.stopBroadcast();
-        
+
         console.log("Deployed new app at:", appAddr);
         console.log("  Owner:", owner);
         console.log("  Disable upgrades:", disableUpgrades);
@@ -183,7 +174,7 @@ contract DeployApp is BaseScript {
         if (composeHash != bytes32(0)) {
             console.log("  Initial compose hash:", vm.toString(composeHash));
         }
-        
+
         return appAddr;
     }
 }
@@ -191,7 +182,7 @@ contract DeployApp is BaseScript {
 contract ShowAppInfo is BaseScript {
     function run() external view {
         require(address(app) != address(0), "APP_CONTRACT_ADDR not set");
-        
+
         console.log("=== App Information ===");
         console.log("App Contract:", address(app));
         console.log("Owner:", app.owner());
@@ -207,35 +198,35 @@ contract BatchKmsSetup is BaseScript {
         bytes32[] memory mrAggregated = vm.envOr("MR_AGGREGATED_LIST", ",", new bytes32[](0));
         bytes32[] memory osImages = vm.envOr("OS_IMAGE_LIST", ",", new bytes32[](0));
         bytes32[] memory devices = vm.envOr("DEVICE_LIST", ",", new bytes32[](0));
-        
+
         vm.startBroadcast();
-        
+
         // Set gateway ID if provided
         if (bytes(gatewayId).length > 0) {
             kms.setGatewayAppId(gatewayId);
             console.log("Set gateway app ID:", gatewayId);
         }
-        
+
         // Add aggregated MRs
-        for (uint i = 0; i < mrAggregated.length; i++) {
+        for (uint256 i = 0; i < mrAggregated.length; i++) {
             kms.addKmsAggregatedMr(mrAggregated[i]);
             console.log("Added MR aggregated:", vm.toString(mrAggregated[i]));
         }
-        
+
         // Add OS images
-        for (uint i = 0; i < osImages.length; i++) {
+        for (uint256 i = 0; i < osImages.length; i++) {
             kms.addOsImageHash(osImages[i]);
             console.log("Added OS image:", vm.toString(osImages[i]));
         }
-        
+
         // Add devices
-        for (uint i = 0; i < devices.length; i++) {
+        for (uint256 i = 0; i < devices.length; i++) {
             kms.addKmsDevice(devices[i]);
             console.log("Added device:", vm.toString(devices[i]));
         }
-        
+
         vm.stopBroadcast();
-        
+
         console.log("\nBatch KMS setup completed!");
     }
 }
@@ -245,11 +236,11 @@ contract BatchKmsSetup is BaseScript {
 contract RemoveOsImage is BaseScript {
     function run() external {
         bytes32 imageHash = vm.envBytes32("OS_IMAGE_HASH");
-        
+
         vm.startBroadcast();
         kms.removeOsImageHash(imageHash);
         vm.stopBroadcast();
-        
+
         console.log("Removed OS image hash:", vm.toString(imageHash));
     }
 }
@@ -257,11 +248,11 @@ contract RemoveOsImage is BaseScript {
 contract RemoveKmsDevice is BaseScript {
     function run() external {
         bytes32 deviceId = vm.envBytes32("DEVICE_ID");
-        
+
         vm.startBroadcast();
         kms.removeKmsDevice(deviceId);
         vm.stopBroadcast();
-        
+
         console.log("Removed KMS device:", vm.toString(deviceId));
     }
 }
@@ -271,11 +262,11 @@ contract RemoveComposeHash is BaseScript {
     function run() external {
         require(address(app) != address(0), "APP_CONTRACT_ADDR not set");
         bytes32 composeHash = vm.envBytes32("COMPOSE_HASH");
-        
+
         vm.startBroadcast();
         app.removeComposeHash(composeHash);
         vm.stopBroadcast();
-        
+
         console.log("Removed compose hash:", vm.toString(composeHash));
     }
 }
@@ -284,11 +275,11 @@ contract RemoveDevice is BaseScript {
     function run() external {
         require(address(app) != address(0), "APP_CONTRACT_ADDR not set");
         bytes32 deviceId = vm.envBytes32("DEVICE_ID");
-        
+
         vm.startBroadcast();
         app.removeDevice(deviceId);
         vm.stopBroadcast();
-        
+
         console.log("Removed device:", vm.toString(deviceId));
     }
 }
@@ -297,11 +288,11 @@ contract RemoveDevice is BaseScript {
 contract RegisterApp is BaseScript {
     function run() external {
         address appAddress = vm.envAddress("APP_ADDRESS");
-        
+
         vm.startBroadcast();
         kms.registerApp(appAddress);
         vm.stopBroadcast();
-        
+
         console.log("Registered app:", appAddress);
     }
 }
@@ -310,11 +301,11 @@ contract RegisterApp is BaseScript {
 contract SetAppImplementation is BaseScript {
     function run() external {
         address appImpl = vm.envAddress("APP_IMPLEMENTATION");
-        
+
         vm.startBroadcast();
         kms.setAppImplementation(appImpl);
         vm.stopBroadcast();
-        
+
         console.log("Set app implementation:", appImpl);
     }
 }
@@ -323,11 +314,11 @@ contract SetAppImplementation is BaseScript {
 contract DisableAppUpgrades is BaseScript {
     function run() external {
         require(address(app) != address(0), "APP_CONTRACT_ADDR not set");
-        
+
         vm.startBroadcast();
         app.disableUpgrades();
         vm.stopBroadcast();
-        
+
         console.log("Disabled upgrades for app:", address(app));
     }
 }
